Update simulate.jl
Added a guard against pushing the same event multiple times. For
example, the same plateau start event will be inserted into the event
queue multiple times, processed multiple times, logged multiple times.
This is due to:
(a) We use the update function to update the event queue, but not to
change the state of the network. Therefore, a segment is only set to
active when the plateau start event is popped off the event queue.
(b) The BinaryMinHeap we use as the event queue will return elements in
order of insertion if they compare equal (because isless() returns
false).
Therefore multiple spikes with the same t will retrigger the update
function and every spike after the first spike that is enough to set a
segment superthreshold will add an additional plateau_start and
plateau_end pair into the event queue.
The same happens when a spike triggers plateaus in two distinct segments
in a neuron. One of the two emitted plateau events for one segment will call the
update function, which will add another plateau for the other segment(s)
since they are still superthreshold but their respective plateau event
has not been processed and they are not set to active.

This fix simply creates a barrier to entering the same event multiple
times into the event queue. The solution is not pretty. However, the
other straightforward alternative does not seem to be much better: mix
the syntax of the update! function as a queue update with the plateau
function -- which makes plateau start events no-ops and confuses
responsibilities of functions.
In the long run, a delicate mini-redesign seems appropriate.